Bot IDs as inputs for generating access tokens
This is a fix for avocado-project tools which use MR. Avocado bot for
manipulation with issue and PR. This tool needs MR. Avocados IDs, which is
stored in Avocado organization as secrets. Unfortunately, the composite actions
cannot use organization secrets, therefore we need to pass them as inputs to
the tool. This commit updates the avocado-project action to use MR. Avocados
IDs as inputs.

Signed-off-by: Jan Richter <jarichte@redhat.com>